fix mock payments issue
